Rename dimension.h
